+++ lib/Util/FileUtils.php
@@ -38,10 +38,7 @@ class FileUtils
 
     public static function renameFile(string $src_path, string $dest_path): bool
     {
-        $src_parsed_url = parse_url($src_path);
-        $dest_parsed_url = parse_url($dest_path);
-
-        if ($src_parsed_url['scheme'] !== $dest_parsed_url['scheme']) {
+        if (!self::isSameScheme($src_path, $dest_path)) {
             // 프로토콜이 다르면 dest로 file copy 후 원본 삭제
             $copy_result = copy($src_path, $dest_path);
 
@@ -53,5 +50,20 @@ class FileUtils
         }
 
         return rename($src_path, $dest_path);
+    }
+
+    public static function isSameScheme(string $src_path, string $dest_path): bool
+    {
+        $src_parsed_url = parse_url($src_path);
+        $dest_parsed_url = parse_url($dest_path);
+
+        return $src_parsed_url['scheme'] === $dest_parsed_url['scheme'];
+    }
+
+    public static function isS3Scheme(string $path): bool
+    {
+        $parsed_url = parse_url($path);
+
+        return preg_match_all('/(s3|S3)/', $parsed_url['scheme']) !== false;
     }
 }